requirements.txt
discord.py
discord.py[voice]
requests
python-dotenv
wikipedia
youtube_dl
asyncio
PyNaCL
ffmpeg
pip